+++ dart/test/default_integrations_test.dart
@@ -47,7 +47,7 @@ void main() {
 
       expect(SentryLevel.fatal, event.level);
 
-      final throwableMechanism = event.throwable as ThrowableMechanism;
+      final throwableMechanism = event.throwableMechanism as ThrowableMechanism;
       expect('isolateError', throwableMechanism.mechanism.type);
       expect(true, throwableMechanism.mechanism.handled);
       expect(throwable, throwableMechanism.throwable);
@@ -103,7 +103,7 @@ void main() {
 
     expect(SentryLevel.fatal, event.level);
 
-    final throwableMechanism = event.throwable as ThrowableMechanism;
+    final throwableMechanism = event.throwableMechanism as ThrowableMechanism;
     expect('runZonedGuarded', throwableMechanism.mechanism.type);
     expect(true, throwableMechanism.mechanism.handled);
     expect(throwable, throwableMechanism.throwable);