+++ CHANGELOG.md
@@ -1,7 +1,14 @@
-# vNext
+# Unreleased
+
+* Fix: Do not append stack trace to the exception if there are no frames
+* Fix: Empty DSN disables the SDK and runs the App
+
+# 4.0.6
 
 * Refactoring: Migrate Sentry Dart to null safety
 * Fix: captureMessage defaults SentryLevel to info
+* Fix: SentryEvent.throwable returns the unwrapped throwable instead of the throwableMechanism
+* Feat: Support enableNativeCrashHandling on iOS
 
 # 4.0.5